[common] bundle all regexps into PatternUtils and move it to 'common'
Also, add a few more comments on the patterns.
